[cutie] Add qtwayland5 back
We use KDE polkit agent because we do not have our own agent (yet).
It is a Qt5 application and we need wayland support for it.
Also, add cutie-camera back. It should get qml-module-cutie (last Qt5 version)
as a dependency.

Moreover, as we use telegram for communication. Let's save some time by having it preinstalled.